requirements.txt
beautifulsoup4==4.14.3
fitz==0.0.1.dev2
google-generativeai==0.8.6
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.4.0
openai==2.14.0
pandas==2.3.3
pdfminer==20191125
pdfminer.six==20251230
PyMuPDF==1.26.7
python-dotenv==1.2.1
pyvi==0.1.1
rank-bm25==0.2.2
selenium==4.39.0
streamlit==1.52.2
webdriver-manager==4.0.2
ipykernel==7.2.0
notebook==7.5.3
loguru==0.7.3
kaggle==1.8.4
jupyterlab-vim==4.1.4
sentence-transformers==5.2.2
imbalanced-learn==0.14.1
pybind11==3.0.2
